fix: Codex レビュー対応 — quality/speed 範囲チェック + rowBytes 安全化 + Case G 追加
- src/root.zig: pict_encode_avif に quality > 100 / speed > 10 の早期 null 返却を追加
- src/c/avif_encode.c: rowBytes を (uint32_t)((size_t)w * (size_t)ch) に変更し 32-bit 乗算オーバーフローを防止
- test/ffi/test.ts: Case G (quality=255・speed=255 の範囲外値が null を返すこと) を追加、合計 8 テストに更新
- docs/CHECKLIST.md: Phase 7C VPS 確認完了・Phase 7B テスト件数更新

Made-with: Cursor

+++ src/c/avif_encode.c
@@ -50,7 +50,7 @@ int pict_avif_encode(
     avifRGBImageSetDefaults(&rgb, image);
     rgb.format   = (channels == 4) ? AVIF_RGB_FORMAT_RGBA : AVIF_RGB_FORMAT_RGB;
     rgb.pixels   = (uint8_t *)pixels;
-    rgb.rowBytes = width * (uint32_t)channels;
+    rgb.rowBytes = (uint32_t)((size_t)width * (size_t)channels);
 
     avifResult conv_result = avifImageRGBToYUV(image, &rgb);
     if (conv_result != AVIF_RESULT_OK) {
